msg/messages --> types,   root/messages --> msgs ...
renamed HaltingState's messages.go file to types.go,      moved root
folder messages.go into msgs subfolder
