+++ msg/id.go
@@ -1,12 +1,14 @@
 package msg
 
 import (
-	"math/rand"
+//"math/rand"
 )
 
 //HyperVisor: processId
 type ProccesId uint64
 
+/*
 func RandProccesId() ProccesId {
 	return (TransportId)(rand.Int63())
 }
+*/
